refactor(ava/advisor): split data-to-advices function and spec generation into sub-files
