+++ packages/ava/src/advisor/advise-pipeline/constants.ts
@@ -1,0 +1,3 @@
+export const DISCRETE_PALETTE_TYPES = ['monochromatic', 'analogous'] as const;
+export const CATEGORICAL_PALETTE_TYPES = ['polychromatic', 'split-complementary', 'triadic', 'tetradic'] as const;
+export const DEFAULT_COLOR = '#678ef2';
