+++ packages/ava/src/advisor/advise-pipeline/index.ts
@@ -1,1 +1,3 @@
-export { advicesForChart } from './advicesForChart';
+export * from './data-to-advices';
+export * from './spec-generator';
+export * from './data-processors';
