+++ packages/ava/src/advisor/advise-pipeline/score-calculator/index.ts
@@ -1,0 +1,2 @@
+export * from './score-rules';
+export * from './compare-advices';
