+++ packages/ava/src/advisor/advise-pipeline/spec-generator/charts/index.ts
@@ -1,0 +1,8 @@
+export * from './area';
+export * from './bar';
+export * from './column';
+export * from './heatmap';
+export * from './histogram';
+export * from './line';
+export * from './pie';
+export * from './scatter';
