+++ packages/ava/src/advisor/advise-pipeline/spec-processors/applyDesignRules.ts
@@ -1,0 +1,19 @@
+import { BasicDataPropertyForAdvice, RuleModule, Specification, DesignRuleModule } from '../../types';
+import { deepMix } from '../../utils';
+
+export function applyDesignRules(
+  chartType: string,
+  dataProps: BasicDataPropertyForAdvice[],
+  ruleBase: Record<string, RuleModule>,
+  chartSpec: Specification
+) {
+  const toCheckRules = Object.values(ruleBase).filter(
+    (rule: RuleModule) =>
+      rule.type === 'DESIGN' && rule.trigger({ dataProps, chartType }) && !ruleBase[rule.id].option?.off
+  );
+  const encodingSpec = toCheckRules.reduce((lastSpec, rule: RuleModule) => {
+    const relatedSpec = (rule as DesignRuleModule).optimizer(dataProps, chartSpec);
+    return deepMix(lastSpec, relatedSpec);
+  }, {});
+  return encodingSpec;
+}
