+++ packages/ava/src/advisor/advise-pipeline/spec-processors/index.ts
@@ -1,0 +1,3 @@
+export * from './applyDesignRules';
+export * from './applySmartColor';
+export * from './applyTheme';
