+++ packages/ava/src/advisor/ruler/utils.ts
@@ -1,10 +1,7 @@
 import { intersects } from '../utils';
-import { DEFAULT_RULE_WEIGHTS } from '../constants';
-import { CHART_IDS } from '../../ckb';
 
-import type { ScoringResultForRule } from '../types';
-import type { ChartKnowledge, LevelOfMeasurement, ChartKnowledgeBase } from '../../ckb';
-import type { BasicDataPropertyForAdvice, ChartRuleModule, Info, RuleModule } from './types';
+import type { ChartKnowledge, LevelOfMeasurement } from '../../ckb';
+import type { BasicDataPropertyForAdvice } from './types';
 
 export function compare(f1: any, f2: any) {
   if (f1.distinct < f2.distinct) {
@@ -28,7 +25,7 @@ export function verifyDataProps(dataPre: ChartKnowledge['dataPres'][number], dat
         lomCount += 1;
       }
     });
-    if (lomCount >= dataPre.minQty && (lomCount <= dataPre.maxQty || dataPre.maxQty === '*')) {
+    if (lomCount >= dataPre.minQty && (dataPre.maxQty === '*' || lomCount <= dataPre.maxQty)) {
       return true;
     }
   }
@@ -38,33 +35,3 @@ export function verifyDataProps(dataPre: ChartKnowledge['dataPres'][number], dat
 export function isUndefined(value: any) {
   return value === undefined;
 }
-
-const defaultWeights = DEFAULT_RULE_WEIGHTS;
-declare type ChartID = (typeof CHART_IDS)[number];
-
-export const computeScore = (
-  chartType: ChartID | string,
-  chartWIKI: ChartKnowledgeBase,
-  ruleBase: Record<string, RuleModule>,
-  ruleType: 'HARD' | 'SOFT',
-  info: Info,
-  log: ScoringResultForRule[]
-) => {
-  // initial score is 1 for HARD rules and 0 for SOFT rules
-  let computedScore = 1;
-  Object.values(ruleBase)
-    .filter((r: RuleModule) => {
-      const weight = r.option?.weight || defaultWeights[r.id] || 1;
-      const extra = r.option?.extra;
-      return r.type === ruleType && r.trigger({ ...info, weight, ...extra, chartType, chartWIKI }) && !r.option?.off;
-    })
-    .forEach((r: RuleModule) => {
-      const weight = r.option?.weight || defaultWeights[r.id] || 1;
-      const extra = r.option?.extra;
-      const base = (r as ChartRuleModule).validator({ ...info, weight, ...extra, chartType, chartWIKI }) as number;
-      const score = weight * base;
-      computedScore *= score;
-      log.push({ phase: 'ADVISE', ruleId: r.id, score, base, weight, ruleType });
-    });
-  return computedScore;
-};
